alx_travel_app/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-filter>=23.0
psycopg2-binary>=2.9.6      # Only if using PostgreSQL
gunicorn>=21.0.0            # For deployment (optional)
